git-svn-id: http://svn.ghostscript.com/ghostscript/branches/icc_work@9875 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/psi/iref.h
@@ -380,7 +380,7 @@ struct ref_s {
     struct tas_s tas;
 
     union v {			/* name the union to keep gdb happy */
-	long intval;
+	int intval;
 	ushort boolval;
 	float realval;
 	ulong saveid;
